+++ pkg/orchestrator/orchestrator_test.go
@@ -17,6 +17,7 @@ package orchestrator
 import (
 	"context"
 	"fmt"
+	"github.com/conduitio/conduit/pkg/processor/procbuiltin"
 	"os"
 	"reflect"
 	"testing"
@@ -86,7 +87,7 @@ func TestPipelineSimple(t *testing.T) {
 	// TODO at the time of writing we don't have a processor for manipulating
 	//  metadata, once we have it we can use it instead of adding our own
 	processor.GlobalBuilderRegistry.MustRegister("removereadat", func(config processor.Config) (processor.Interface, error) {
-		return processor.NewFuncWrapper(func(ctx context.Context, r record.Record) (record.Record, error) {
+		return procbuiltin.NewFuncWrapper(func(ctx context.Context, r record.Record) (record.Record, error) {
 			delete(r.Metadata, record.MetadataReadAt) // read at is different every time, remove it
 			return r, nil
 		}), nil
