+++ pkg/processor/procbuiltin/func_wrapper_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package processor
+package procbuiltin
 
 import (
 	"context"
