+++ pkg/processor/procbuiltin/maskfield.go
@@ -69,7 +69,7 @@ func maskField(
 	}
 	replacement = config.Settings[maskFieldConfigReplacement]
 
-	return processor.NewFuncWrapper(func(_ context.Context, r record.Record) (record.Record, error) {
+	return NewFuncWrapper(func(_ context.Context, r record.Record) (record.Record, error) {
 		data := getSetter.Get(r)
 
 		switch d := data.(type) {
