+++ pkg/processor/procbuiltin/replacefield.go
@@ -118,7 +118,7 @@ func replaceField(
 		}
 	}
 
-	return processor.NewFuncWrapper(func(_ context.Context, r record.Record) (record.Record, error) {
+	return NewFuncWrapper(func(_ context.Context, r record.Record) (record.Record, error) {
 		data := getSetter.Get(r)
 
 		switch d := data.(type) {
