+++ pkg/provisioning/service_test.go
@@ -16,6 +16,7 @@ package provisioning
 
 import (
 	"context"
+	"github.com/conduitio/conduit/pkg/processor/procbuiltin"
 	"os"
 	"sort"
 	"testing"
@@ -684,7 +685,7 @@ func TestProvision_IntegrationTestServices(t *testing.T) {
 	// TODO at the time of writing we don't have a processor for manipulating
 	//  metadata, once we have it we can use it instead of adding our own
 	processor.GlobalBuilderRegistry.MustRegister("removereadat", func(config processor.Config) (processor.Interface, error) {
-		return processor.NewFuncWrapper(func(ctx context.Context, r record.Record) (record.Record, error) {
+		return procbuiltin.NewFuncWrapper(func(ctx context.Context, r record.Record) (record.Record, error) {
 			delete(r.Metadata, record.MetadataReadAt) // read at is different every time, remove it
 			return r, nil
 		}), nil
